Manually execute fieldChange before submit
This way choosing a value and clicking submit before the
debounce-Timeout runs out will not cause issues anymore.
During the click of the submit button the process will check for any
fields that have a pending change and execute it immediately.

Before it was possible to choose an address and very quickly click
submit. The sequence would be the following:
1. select value
2. click submit
3. send submit request to backend
4. customer is redirected to estimator, where he will find the field
from step 1 unfilled
5. debouncetimeout expires -> send chosen value to backend (already too
late)

Also set the debounceWaittime for a google place id to 200 ms. It is
highly unlikely that this is going to change, so we prefer sending it
quickly.